requirements.txt
torch>=2.1.0
tqdm
numpy
transformer_lens
huggingface_hub
bitsandbytes
wandb
matplotlib
seaborn
scikit-learn
hydra-core